+++ internal/provider/monitor/data_source_acc_test.go
@@ -46,12 +46,12 @@ data "uptimerobot_monitor" "by_name" {
 func testAccCheckMonitorVisibleInList(name string) resource.TestCheckFunc {
 	return func(_ *terraform.State) error {
 		apiClient := provideracctest.APIClient()
-		ctx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
+		ctx, cancel := context.WithTimeout(context.Background(), 10*time.Minute)
 		defer cancel()
 
 		var lastGetMonitorsErr error
 		for {
-			monitors, err := apiClient.GetMonitors(ctx)
+			monitors, err := apiClient.GetMonitorsByName(ctx, name)
 			if err != nil {
 				lastGetMonitorsErr = err
 			} else {
@@ -65,10 +65,10 @@ func testAccCheckMonitorVisibleInList(name string) resource.TestCheckFunc {
 			select {
 			case <-ctx.Done():
 				if lastGetMonitorsErr != nil {
-					return fmt.Errorf("monitor %q was not visible in list endpoint before ctx.Done; last apiClient.GetMonitors error: %v: %w", name, lastGetMonitorsErr, ctx.Err())
+					return fmt.Errorf("monitor %q was not visible in list endpoint before ctx.Done; last apiClient.GetMonitorsByName error: %v: %w", name, lastGetMonitorsErr, ctx.Err())
 				}
 				return fmt.Errorf("monitor %q was not visible in list endpoint before timeout: %w", name, ctx.Err())
-			case <-time.After(2 * time.Second):
+			case <-time.After(5 * time.Second):
 			}
 		}
 	}